+++ StickyHeaderTabController/Classes/Views/StickyHeaderHeroView.swift
@@ -15,4 +15,10 @@ public protocol StickyHeaderHeroViewDelegate: class {
 
 open class StickyHeaderHeroView: UIView {
 
+    /// A kind of hacky way of letting view specify its height.
+    /// TODO: refactor in terms of preferredContentSize?
+    open var heroHeight: CGFloat {
+        return 200.0
+    }
+
 }