+++ StickyHeaderTabController/Classes/Views/StickyHeaderView.swift
@@ -17,5 +17,16 @@ open class StickyHeaderView: UIView {
     // MARK: - Public Properties
 
     public weak var delegate: StickyHeaderViewDelegate?
+
+    /// A kind of hacky way of letting view specify its height.
+    /// TODO: refactor in terms of preferredContentSize?
+    open var headerHeight: CGFloat {
+        return 170.0
+    }
+
+    /// The minimum header height (while pinned in sticky top position)
+    open var pinnedHeight: CGFloat {
+        return 60.0
+    }
     
 }